Moving the extractCodeBlockLanguage function to shared/markdown.ts and added test cases for this function
